day-06.1.sch
AI 기능 스크립트 제작 시작
내용 - 기본적인 틀 제작
첫 목표 - 동료캐릭터의 이동

+++ Assets/#Scripts/Movable/MovableBase.cs
@@ -11,9 +11,9 @@ public abstract class MovableBase : MonoBehaviour
     protected Enums.State state = Enums.State.Default; // 상태(FSM을 위한)
 
     // 추상 메서드들
-    abstract public int Attack();      // 공격관련 메서드
-    abstract public int GetDamage();   // 피격관련 메서드
-    abstract public int GetHeal();     // 체력회복 메서드
+    abstract public float Attack();      // 공격관련 메서드
+    abstract public float GetDamage();   // 피격관련 메서드
+    abstract public float GetHeal();     // 체력회복 메서드
     abstract public void Move();        // 이동관련 메서드
 
     protected virtual void Start()
