+++ Assets/#Scripts/Movable/Playable/PlayeableCharacterBase.cs
@@ -10,17 +10,17 @@ public class PlayeableCharacterBase : MovableBase
         
     }
 
-    public override int Attack()
+    public override float Attack()
     {
         throw new System.NotImplementedException();
     }
 
-    public override int GetDamage()
+    public override float GetDamage()
     {
         throw new System.NotImplementedException();
     }
 
-    public override int GetHeal()
+    public override float GetHeal()
     {
         throw new System.NotImplementedException();
     }
